requirements.txt
flask
flask_assets
pyrebase4
firebase_admin